backend/requirements.txt
fastapi[all]==0.105.0
SQLAlchemy==2.0.23
python-jose==3.3.0
bcrypt==4.1.2
passlib==1.7.4
PyMySQL==1.1.1
pillow==10.2.0
jinja2==3.1.3
psycopg2==2.9.9
mysqlclient==2.2.4
